Disable -march=native which breaks builds
Replace it with x86-64-v3 which all users of this CPU support.

When copr built this with a big server, it built it with AVX-512.
Sadly, Intel does not currently support AVX-512 on its consumer
platforms with NPUs.
This would lead to SIGILLs when attempting to test this on hardware.

(It's not great since Fedora packages should support all arches, but
this package uses AVX2 intriniscs and I'd like to get it at least
building before doing deeper fixes for older platforms that don't
even have an Intel NPU anyway.)

Signed-off-by: Alexander F. Lent <lx@xanderlent.com>